wrapper: parallel phase A boot, per-probe durations, split into feature files
Start chromium alongside xorg/xvfb so chromium-launcher's preamble overlaps
display startup; chromium-launcher waits internally for the X server.
Final ready log reports per-probe (cdp/chromedriver/forward-proxy/...)
durations so the gating probe is visible. main.go split into feature files
(supervisord, probes, display, chromium, envoy, system).
